+++ blocksuite/blocks/src/effects.ts
@@ -16,7 +16,6 @@ import { effects as blockRootEffects } from '@blocksuite/affine-block-root/effec
 import { effects as blockSurfaceEffects } from '@blocksuite/affine-block-surface/effects';
 import { effects as blockSurfaceRefEffects } from '@blocksuite/affine-block-surface-ref/effects';
 import { effects as blockTableEffects } from '@blocksuite/affine-block-table/effects';
-import { effects as componentAiItemEffects } from '@blocksuite/affine-components/ai-item';
 import { BlockSelection } from '@blocksuite/affine-components/block-selection';
 import { BlockZeroWidth } from '@blocksuite/affine-components/block-zero-width';
 import { effects as componentCaptionEffects } from '@blocksuite/affine-components/caption';
@@ -82,7 +81,6 @@ export function effects() {
   componentToolbarEffects();
   componentDropIndicatorEffects();
   componentToggleButtonEffects();
-  componentAiItemEffects();
   componentColorPickerEffects();
   componentEmbedCardModalEffects();
   componentDocTitleEffects();
