+++ packages/frontend/core/src/blocksuite/presets/ai/_common/components/ai-item/ai-item.ts
@@ -2,12 +2,12 @@ import {
   EditorHost,
   PropTypes,
   requiredProperties,
-} from '@blocksuite/block-std';
-import { WithDisposable } from '@blocksuite/global/utils';
+} from '@blocksuite/affine/block-std';
+import { ArrowRightIcon, EnterIcon } from '@blocksuite/affine/blocks';
+import { WithDisposable } from '@blocksuite/affine/global/utils';
 import { css, html, LitElement, nothing } from 'lit';
 import { property, query } from 'lit/decorators.js';
 
-import { ArrowRightIcon, EnterIcon } from '../icons';
 import { menuItemStyles } from './styles';
 import type { AIItemConfig } from './types';
