+++ packages/frontend/core/src/blocksuite/presets/ai/_common/components/ai-item/const.ts
