+++ packages/frontend/core/src/blocksuite/presets/ai/peek-view/chat-block-input.ts
@@ -1,9 +1,5 @@
 import type { EditorHost } from '@blocksuite/affine/block-std';
-import {
-  type AIError,
-  openFileOrFiles,
-  unsafeCSSVarV2,
-} from '@blocksuite/affine/blocks';
+import { openFileOrFiles, unsafeCSSVarV2 } from '@blocksuite/affine/blocks';
 import { SignalWatcher } from '@blocksuite/affine/global/utils';
 import { ImageIcon, PublishIcon } from '@blocksuite/icons/lit';
 import { css, html, LitElement, nothing } from 'lit';
@@ -11,6 +7,7 @@ import { property, query, state } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 
 import type { ChatMessage } from '../../../blocks';
+import type { AIError } from '../_common/components/ai-item/types';
 import { ChatAbortIcon, ChatClearIcon, ChatSendIcon } from '../_common/icons';
 import type { AINetworkSearchConfig } from '../chat-panel/chat-config';
 import {
